Implement a simple, "HTTP" server to initiate profiling
This implements a basic, not fully compliant HTTP server that will respond
to a single endpoint, "/profile", on a random free port.

The port will be logged as well as indicated via the name of a tempfile in a
specific location.

This will allow an external process to pull ruby profiles of a specified type
and duration for the target process.