Update profile server to respond with a conflict if already profiling
It is possible that a per-request profile could coincide with a request to
profile in the background via the profile server. This edge case is currently
unhandled, as we don't check the return code of `AppProfiler.start`.

This adds coverage for this edge case, so that it will respond with an HTTP
conflict, as there is already a profile being taken.