Use HTML5 sanitizer to serve huge payloads through Speedscope remote viewer
When we run the app profiler on Shopify core's boot, it is frequent that
the generated profile JSON file exceeds 10MB.

`Rails::Html::Sanitizer` is unable to process such huge payloads
(because libxml2 limits text nodes to 10MB by default?).

The `rails-html-sanitizer` introduced HTML5 parsing in its 1.6.0
version, which lifts the text node size limit.

This change intends to use the new HTML5 parser when it's available
to sanitize the Speedscope remote viewer's HTML, allowing bigger
payloads to be served successfully.