Allow modular profiling backends, support vernier
This adds support for the vernier profiler by replacing the notion of a
"profiler" with a profiler "backend". A basic interface based on the current
profiler contract is abstracted, which is implemented by both vernier and
stackprof.

Stackprof is the default profiler backend by default. Vernier is only supported
if the user explicitly adds it in their gemfile, and it satisfies the version
constraints within app_profiler. Vernier code is only autoloaded on first use,
and should not interfere with existing profiler operations via stackprof.